pyproject.toml
[project]
name = "translate-dir-lib"
version = "0.1.0"
description = "A library for automated directory translation"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.16.1",
    "httpx>=0.28.1",
    "pydantic>=2.11.5",
    "typer>=0.16.0",
]

[project.scripts]
translate-dir = "cli:app"
